Renamed some references to import-blockchain-data for consistency with dogeboxd
