+++ src/api/mocks.js
@@ -40,7 +40,7 @@ import {
   getResponse as checkForUpdates,
   postResponse as commenceUpdate,
 } from "./system/updates.mocks.js";
-import { postResponse as importBlockchain } from "./system/import-blockchain.mocks.js";
+import { postResponse as importBlockchain } from "./system/import-blockchain-data.mocks.js";
 
 import {
   getSSHPublicKeysResponse,
