+++ src/api/system/import-blockchain-data.js
@@ -3,11 +3,11 @@ import { store } from '/state/store.js';
 
 import { 
   postResponse,
-} from './import-blockchain.mocks.js';
+} from './import-blockchain-data.mocks.js';
 
 const client = new ApiClient(store.networkContext.apiBaseUrl);
 
 export async function importBlockchain() {
-  const res = await client.post(`/system/import-blockchain`, {}, { mock: postResponse });
+  const res = await client.post(`/system/import-blockchain-data`, {}, { mock: postResponse });
   return res;
 }
